Add a zeroInitialise() method to the Portable objects and collections
Add a zeroInitialise() method to the PortableObject and PortableCollection
templates that erases the data in the Buffer by writing zeros (bytes
containing '\0') to it.

On the host, zeroInitialise() can be called without arguments to execute
immediately, or with a queue to be run asynchronously.

On the device, zeroInitialise() requires a queue and can only be run
asynchronously.